backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
bcrypt==4.0.1
python-dateutil==2.8.2

# AI Providers (install yang kamu butuhkan)
openai>=1.0.0
# anthropic>=0.18.0
# google-generativeai>=0.3.0
# groq>=0.4.0